+++ Src/Actor.cs
@@ -20,6 +20,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Management.Automation;
 using MongoDB.Bson;
+using MongoDB.Bson.IO;
 using MongoDB.Bson.Serialization;
 using MongoDB.Driver;
 using MongoDB.Driver.Builders;
@@ -31,7 +32,10 @@ namespace Mdbc
 		public const string ServerVariable = "Server";
 		public const string DatabaseVariable = "Database";
 		public const string CollectionVariable = "Collection";
+
 		static bool _registered;
+
+		// This method should be called as soon as possible. Multiple calls are allowed.
 		public static void Register()
 		{
 			if (_registered)
@@ -45,6 +49,17 @@ namespace Mdbc
 			BsonSerializer.RegisterSerializer(typeof(PSObject), new PSObjectSerializer());
 
 			BsonTypeMapper.RegisterCustomTypeMapper(typeof(PSObject), new PSObjectTypeMapper());
+		}
+		// NB: Change of the global defaults affects ToString().
+		// `Strict` (like mongoexport) is tempting, other tools may read it.
+		// But `Strict` is not suitable for reading and searching. Also, it
+		// looses number types (numbers are double). `Shell` (default) keeps
+		// types and readable for _id, dates, GUID. So let's use `Shell` not
+		// directly but via defaults. If needed defaults can be changed by a
+		// user.
+		public static JsonWriterSettings DefaultJsonWriterSettings
+		{
+			get { return JsonWriterSettings.Defaults; }
 		}
 		public static object BaseObject(object value)
 		{